% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzedata.R
\name{busiest_routes}
\alias{busiest_routes}
\title{Determines which routes are the busiest}
\usage{
busiest_routes(dataframe, startcol, endcol)
}
\arguments{
\item{dataframe}{dataframe to be used}

\item{startcol}{column for the starting location of the flight}

\item{endcol}{column for the ending location of the flight}
}
\value{
dataset containing market shares by airline and airport
}
\description{
Determines which routes are the busiest
}
