% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{data_clean}
\alias{data_clean}
\title{Cleans the datafiles to combine them into one}
\usage{
data_clean(datafile, cityfile, carrierfile)
}
\arguments{
\item{datafile}{main datafile with flight info}

\item{cityfile}{data file with information on cities}

\item{carrierfile}{data file with info on airline carrier}
}
\value{
a joined and cleaned datafile
}
\description{
Cleans the datafiles to combine them into one
}
