% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzedata.R
\name{market_shares}
\alias{market_shares}
\title{Determines market shares for airlines in airports}
\usage{
market_shares(dataframe, carrier, origin)
}
\arguments{
\item{dataframe}{dataframe to be used}

\item{carrier}{column for airline carrier}

\item{origin}{column for city/airport or origin}
}
\value{
dataset containing percent of passengers in each airport for an airline
}
\description{
Determines market shares for airlines in airports
}
